Refactored UI actions to separate module "renderUI.mjs" - proved to be quite difficult to separate todoList from the UI. Am not certain this is even the best structure. Had to comment out one test (to revisit)
