+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootswatch/4.4.1/sketchy/bootstrap.min.css"/>
     <style>
       body {
-        background-color:powderblue;
+        background-color:black; /*powderblue;*/
      /*   font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;*/
         font-size: x-large;
       }
@@ -40,12 +40,13 @@
 
   <body>
     <div style="padding: 5vw;">
-      <h1>My Todo List:</h1>
+      <h1 style="color:white">My Todo List:</h1>
       <div>
         <button type="button" id="btn-toggle-all" class="btn btn-info"  >TOGGLE</BUTTON>      
         <input type="text" placeholder="e.g. buy some milk" id="inputTodo" size="50"/>
-        <button type="button" class="btn btn-warning" id="btn-add-todo">ADD</button>
+        <button type="button" class="btn btn-success" id="btn-add-todo">ADD</button>
       </div> 
+      <br/>
       <div id="id-todos" class="matrix-chalk-purple" style="height:80vh;"><!--Placeholder-->
       </div>
       <!-- NOT USED:
